Add files via upload
Alpha version of BlackJack created.
I will be developing other card games in the mean time, but this will be black jack for now. Lots of editing needed to be done, and lots of refactoring. I am currently researching development for the dealer, which will hopefully employ some AI stuff.